refactor(web): split auth/oauth/spotify routes into routers (Phase 6b, Group 3)
Moves the highest-risk remaining closures out of create_app(): Google/GitHub
login OAuth and Spotify OAuth+API, into ev/interfaces/web/routes/oauth_google.py,
oauth_github.py, and spotify.py, following the same build_router(ctx) pattern
as prior groups. The shared OAuth CSRF-state set and login HTML helpers move
to WebContext (ctx.oauth_states/base_url/login_ok_html/login_denied_html) so
all three OAuth flows keep validating each other's state values exactly as
before, without duplicating that shared state per-router.
