+++ ev/interfaces/web/app.py
@@ -25,7 +25,8 @@ from .context import WebContext
 from .frontend import _DEFAULT_FOLDERS, _FAVICON, _SERVICE_WORKER, _icon_png, _PAGE
 from .routes import (
     activity, backup, connectors, expenses, facts, habits, journal, keys,
-    links, notifications, pages, push, recurring, reminders, tasks, watches,
+    links, notifications, oauth_github, oauth_google, pages, push, recurring,
+    reminders, spotify, tasks, watches,
 )
 
 log = logging.getLogger("ev.web")
@@ -59,7 +60,8 @@ def create_app(config: Config, brain: Brain | None = None):
 
     for _router_mod in (
         activity, backup, connectors, expenses, facts, habits, journal, keys,
-        links, notifications, pages, push, recurring, reminders, tasks, watches,
+        links, notifications, oauth_github, oauth_google, pages, push,
+        recurring, reminders, spotify, tasks, watches,
     ):
         app.include_router(_router_mod.build_router(ctx))
 
@@ -899,293 +901,6 @@ def create_app(config: Config, brain: Brain | None = None):
         ok = memory.set_local_confirm_status(owner, cid, "rejected")
         return {"ok": ok}
 
-    # --- Spotify OAuth + Web API (Premium: read playlists + control playback) --
-    import time as _time
-
-    def _sp_redirect(request: Request) -> str:
-        return _sp.norm_redirect(config.web_base_url or str(request.base_url))
-
-    def _sp_save(tokens: dict):
-        memory.set_setting("spotify_tokens", json.dumps(tokens))
-
-    def _sp_tokens() -> dict:
-        try:
-            return json.loads(memory.get_setting("spotify_tokens") or "{}")
-        except (ValueError, TypeError):
-            return {}
-
-    def _sp_access() -> str | None:
-        """A valid access token, refreshing with the refresh_token if expired."""
-        import httpx
-        t = _sp_tokens()
-        if not t.get("refresh"):
-            return None
-        if t.get("access") and t.get("exp", 0) > _time.time() + 30:
-            return t["access"]
-        try:
-            data = {"grant_type": "refresh_token", "refresh_token": t["refresh"],
-                    "client_id": config.spotify_client_id}
-            if config.spotify_client_secret:
-                data["client_secret"] = config.spotify_client_secret
-            r = httpx.post("https://accounts.spotify.com/api/token",
-                           data=data, timeout=15).json()
-        except Exception:
-            return None
-        if not r.get("access_token"):
-            return None
-        t["access"] = r["access_token"]
-        t["exp"] = _time.time() + int(r.get("expires_in", 3600))
-        if r.get("refresh_token"):
-            t["refresh"] = r["refresh_token"]
-        _sp_save(t)
-        return t["access"]
-
-    def _sp_api(method: str, path: str, token: str, **kw):
-        import httpx
-        url = path if path.startswith("http") else ("https://api.spotify.com/v1" + path)
-        return httpx.request(method, url, headers={"Authorization": "Bearer " + token},
-                             timeout=15, **kw)
-
-    @app.get("/spotify/connect")
-    async def spotify_connect(request: Request):
-        if not config.spotify_client_id:
-            return _login_denied_html("Configure o Spotify Client ID em Chaves de API.")
-        state = _secrets.token_urlsafe(16); _oauth_states.add(state)
-        verifier, challenge = _sp.pkce_pair(); _sp_pkce[state] = verifier
-        return RedirectResponse(_sp.auth_url(
-            config.spotify_client_id, _sp_redirect(request), state, challenge))
-
-    @app.get("/spotify/callback")
-    async def spotify_callback(request: Request):
-        import httpx
-        code = request.query_params.get("code")
-        state = request.query_params.get("state")
-        if not code or state not in _oauth_states:
-            return _login_denied_html("Sessão do Spotify inválida. Tente de novo.")
-        _oauth_states.discard(state)
-        redirect = _sp_redirect(request)
-        verifier = _sp_pkce.pop(state, "")
-
-        def _work():
-            data = {"grant_type": "authorization_code", "code": code,
-                    "redirect_uri": redirect, "client_id": config.spotify_client_id}
-            if verifier:  # PKCE — sem secret
-                data["code_verifier"] = verifier
-            elif config.spotify_client_secret:
-                data["client_secret"] = config.spotify_client_secret
-            return httpx.post("https://accounts.spotify.com/api/token",
-                              data=data, timeout=15).json()
-        try:
-            r = await asyncio.to_thread(_work)
-        except Exception as exc:
-            return _login_denied_html(f"Falha ao conectar o Spotify: {exc}")
-        if not r.get("refresh_token"):
-            return _login_denied_html("Spotify não devolveu o token. Confira o Client Secret e a Redirect URI.")
-        _sp_save({"access": r.get("access_token"), "refresh": r["refresh_token"],
-                  "exp": _time.time() + int(r.get("expires_in", 3600))})
-        return HTMLResponse("<script>location.href='/'</script>Spotify conectado! Redirecionando…")
-
-    @app.get("/api/spotify/status")
-    async def spotify_status(request: Request):
-        _check(request.headers.get("authorization"))
-        return {"configured": bool(config.spotify_client_id),
-                "connected": bool(_sp_tokens().get("refresh")),
-                "redirect_uri": _sp_redirect(request)}
-
-    @app.post("/api/spotify/disconnect")
-    async def spotify_disconnect(request: Request):
-        _check(request.headers.get("authorization"))
-        memory.set_setting("spotify_tokens", "")
-        return {"ok": True}
-
-    @app.get("/api/spotify/token")
-    async def spotify_token(request: Request):
-        # fresh access token for the Web Playback SDK (client-side)
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        return {"token": tok}
-
-    @app.get("/api/spotify/playlists")
-    async def spotify_playlists(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        def _work():
-            r = _sp_api("GET", "/me/playlists?limit=50", tok)
-            return r.json() if r.status_code == 200 else {}
-        data = await asyncio.to_thread(_work)
-        items = [{"name": p.get("name"), "uri": p.get("uri"), "id": (p.get("id") or ""),
-                  # a API passou a devolver a contagem em items.total (antes tracks.total)
-                  "tracks": ((p.get("tracks") or p.get("items") or {}).get("total", 0))}
-                 for p in (data.get("items") or []) if p]
-        return {"items": items}
-
-    @app.post("/api/spotify/play")
-    async def spotify_play(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        d = await _body(request)
-        body, dev = {}, d.get("device_id")
-        if d.get("uris"):
-            body["uris"] = d["uris"]
-        elif d.get("uri"):
-            u = d["uri"]
-            if u.startswith("spotify:track:"):
-                body["uris"] = [u]           # a single track
-            else:
-                body["context_uri"] = u      # playlist/album/artist
-        if d.get("query"):                   # search then play the top track
-            tk = await asyncio.to_thread(_sp.first_track_uri, tok, d["query"])
-            if tk:
-                body = {"uris": [tk]}
-        def _work():
-            path = "/me/player/play" + (f"?device_id={dev}" if dev else "")
-            r = _sp_api("PUT", path, tok, json=body)
-            return r.status_code
-        sc = await asyncio.to_thread(_work)
-        if sc == 404:
-            return {"ok": False, "error": "nenhum dispositivo ativo — abra o Spotify ou use o player da E.V."}
-        return {"ok": sc in (200, 202, 204)}
-
-    @app.post("/api/spotify/transfer")
-    async def spotify_transfer(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        dev = (await _body(request)).get("device_id")
-        if not dev:
-            raise HTTPException(status_code=400, detail="sem device da E.V.")
-        # move a reprodução pro device da E.V. (toca in-page -> controles do SO)
-        sc = await asyncio.to_thread(lambda: _sp_api(
-            "PUT", "/me/player", tok, json={"device_ids": [dev], "play": True}).status_code)
-        return {"ok": sc in (200, 202, 204)}
-
-    @app.get("/api/spotify/search")
-    async def spotify_search(request: Request):
-        _check(request.headers.get("authorization"))
-        q = (request.query_params.get("q") or "").strip()
-        if not q:
-            return {"items": []}
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        return {"items": await asyncio.to_thread(_sp.search_tracks, tok, q)}
-
-    @app.get("/api/spotify/current")
-    async def spotify_current(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            return {"track": ""}
-        return {"track": await asyncio.to_thread(_sp.current_track, tok)}
-
-    @app.get("/api/spotify/nowplaying")
-    async def spotify_nowplaying(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            return {"playing": False, "connected": False}
-        def _work():
-            r = _sp_api("GET", "/me/player", tok)
-            if r.status_code != 200:
-                return {"playing": False, "connected": True}
-            d = r.json() or {}
-            it = d.get("item") or {}
-            imgs = (it.get("album") or {}).get("images") or []
-            tid = it.get("id")
-            liked = False
-            if tid:
-                try:
-                    lr = _sp_api("GET", "/me/tracks/contains?ids=" + tid, tok)
-                    liked = bool((lr.json() or [False])[0]) if lr.status_code == 200 else False
-                except Exception:
-                    pass
-            dev = d.get("device") or {}
-            return {"connected": True, "playing": bool(d.get("is_playing")),
-                    "name": it.get("name"), "id": tid, "liked": liked,
-                    "artists": ", ".join(a.get("name", "") for a in (it.get("artists") or [])),
-                    "image": (imgs[0].get("url") if imgs else ""),
-                    "progress": d.get("progress_ms", 0), "duration": it.get("duration_ms", 0),
-                    "device": dev.get("name", ""), "volume": dev.get("volume_percent", 50)}
-        return await asyncio.to_thread(_work)
-
-    @app.post("/api/spotify/like")
-    async def spotify_like(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        d = await _body(request)
-        tid = d.get("id")
-        if not tid:
-            raise HTTPException(status_code=400, detail="sem id")
-        method = "PUT" if d.get("on") else "DELETE"
-        sc = await asyncio.to_thread(lambda: _sp_api(method, "/me/tracks?ids=" + tid, tok).status_code)
-        return {"ok": sc in (200, 201, 202, 204)}
-
-    @app.get("/api/spotify/devices")
-    async def spotify_devices(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            return {"items": []}
-        def _work():
-            r = _sp_api("GET", "/me/player/devices", tok)
-            return (r.json().get("devices") or []) if r.status_code == 200 else []
-        ds = await asyncio.to_thread(_work)
-        return {"items": [{"id": x.get("id"), "name": x.get("name"),
-                           "active": x.get("is_active")} for x in ds]}
-
-    @app.post("/api/spotify/volume")
-    async def spotify_volume(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        pct = max(0, min(100, int((await _body(request)).get("percent") or 0)))
-        sc = await asyncio.to_thread(lambda: _sp_api("PUT", f"/me/player/volume?volume_percent={pct}", tok).status_code)
-        return {"ok": sc in (200, 202, 204)}
-
-    @app.post("/api/spotify/queue")
-    async def spotify_queue(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        uri = (await _body(request)).get("uri")
-        if not uri:
-            raise HTTPException(status_code=400, detail="sem uri")
-        import urllib.parse as _up
-        sc = await asyncio.to_thread(lambda: _sp_api("POST", "/me/player/queue?uri=" + _up.quote(uri), tok).status_code)
-        return {"ok": sc in (200, 202, 204)}
-
-    @app.post("/api/spotify/control")
-    async def spotify_control(request: Request):
-        _check(request.headers.get("authorization"))
-        tok = await asyncio.to_thread(_sp_access)
-        if not tok:
-            raise HTTPException(status_code=400, detail="não conectado")
-        d = await _body(request)
-        act = d.get("action")
-        if act == "seek":   # pular pra uma posição (ms)
-            ms = max(0, int(d.get("ms") or 0))
-            sc = await asyncio.to_thread(
-                lambda: _sp_api("PUT", f"/me/player/seek?position_ms={ms}", tok).status_code)
-            return {"ok": sc in (200, 202, 204)}
-        M = {"pause": ("PUT", "/me/player/pause"), "resume": ("PUT", "/me/player/play"),
-             "next": ("POST", "/me/player/next"), "prev": ("POST", "/me/player/previous")}
-        if act not in M:
-            raise HTTPException(status_code=400, detail="ação inválida")
-        method, path = M[act]
-        sc = await asyncio.to_thread(lambda: _sp_api(method, path, tok).status_code)
-        return {"ok": sc in (200, 202, 204)}
-
     # --- Subscriptions / Budgets / Watches CRUD ----------------------------
     # (recurring/watches CRUD, including their reunified /update, moved to
     # .routes/recurring.py and .routes/watches.py — Phase 6b, Group 2)
@@ -1795,127 +1510,6 @@ def create_app(config: Config, brain: Brain | None = None):
             return {"ok": True}
         except Exception as exc:
             return {"ok": False, "msg": str(exc)}
-
-    # --- OAuth login (Google / GitHub) -------------------------------------
-    import secrets as _secrets
-    from fastapi.responses import RedirectResponse
-    _oauth_states: set[str] = set()
-    _sp_pkce: dict = {}   # state -> code_verifier (Spotify PKCE, sem secret)
-
-    def _base_url(request: Request) -> str:
-        return config.web_base_url or str(request.base_url).rstrip("/")
-
-    def _login_ok_html() -> HTMLResponse:
-        # Passed the identity check: hand the app token to this browser and enter.
-        return HTMLResponse(
-            "<!doctype html><meta charset=utf-8><title>E.V.</title>"
-            "<script>localStorage.setItem('ev_token'," + json.dumps(config.web_token)
-            + ");location.replace('/');</script>"
-            "<p style='font:14px system-ui;color:#888;padding:24px'>Entrando…</p>")
-
-    def _login_denied_html(msg: str) -> HTMLResponse:
-        return HTMLResponse(
-            "<!doctype html><meta charset=utf-8><title>E.V.</title>"
-            "<div style='font:15px system-ui;color:#d6e9fb;background:#04070c;"
-            "height:100vh;display:flex;flex-direction:column;align-items:center;"
-            "justify-content:center;gap:16px;text-align:center;padding:24px'>"
-            "<p>" + msg + "</p><a href='/' style='color:#8ab4f8'>voltar</a></div>",
-            status_code=403)
-
-    @app.get("/auth/google")
-    async def auth_google(request: Request):
-        if not config.google_login_client:
-            return _login_denied_html("Login com Google não configurado.")
-        from urllib.parse import urlencode
-        state = _secrets.token_urlsafe(16); _oauth_states.add(state)
-        params = urlencode({
-            "client_id": config.google_login_client,
-            "redirect_uri": _base_url(request) + "/auth/google/callback",
-            "response_type": "code", "scope": "openid email profile",
-            "access_type": "online", "state": state, "prompt": "select_account",
-        })
-        return RedirectResponse("https://accounts.google.com/o/oauth2/v2/auth?" + params)
-
-    @app.get("/auth/google/callback")
-    async def auth_google_cb(request: Request):
-        import httpx
-        code = request.query_params.get("code")
-        state = request.query_params.get("state")
-        if not code or state not in _oauth_states:
-            return _login_denied_html("Sessão de login inválida. Tente de novo.")
-        _oauth_states.discard(state)
-        redirect = _base_url(request) + "/auth/google/callback"
-
-        def _work():
-            tok = httpx.post("https://oauth2.googleapis.com/token", data={
-                "code": code, "client_id": config.google_login_client,
-                "client_secret": config.google_login_secret,
-                "redirect_uri": redirect, "grant_type": "authorization_code",
-            }, timeout=15).json()
-            at = tok.get("access_token")
-            info = httpx.get("https://www.googleapis.com/oauth2/v2/userinfo",
-                             headers={"Authorization": "Bearer " + (at or "")},
-                             timeout=15).json()
-            return (info.get("email") or "").lower()
-        try:
-            email = await asyncio.to_thread(_work)
-        except Exception as exc:
-            return _login_denied_html(f"Falha no login Google: {exc}")
-        if not email:
-            return _login_denied_html("Não consegui obter seu email do Google.")
-        allowed = memory.get_setting("login_google_email")
-        if not allowed:
-            memory.set_setting("login_google_email", email)  # pin the first login
-        elif email != allowed:
-            return _login_denied_html("Esta conta Google não tem acesso a esta E.V.")
-        return _login_ok_html()
-
-    @app.get("/auth/github")
-    async def auth_github(request: Request):
-        if not config.github_login_client:
-            return _login_denied_html("Login com GitHub não configurado.")
-        from urllib.parse import urlencode
-        state = _secrets.token_urlsafe(16); _oauth_states.add(state)
-        params = urlencode({
-            "client_id": config.github_login_client,
-            "redirect_uri": _base_url(request) + "/auth/github/callback",
-            "scope": "read:user", "state": state,
-        })
-        return RedirectResponse("https://github.com/login/oauth/authorize?" + params)
-
-    @app.get("/auth/github/callback")
-    async def auth_github_cb(request: Request):
-        import httpx
-        code = request.query_params.get("code")
-        state = request.query_params.get("state")
-        if not code or state not in _oauth_states:
-            return _login_denied_html("Sessão de login inválida. Tente de novo.")
-        _oauth_states.discard(state)
-        redirect = _base_url(request) + "/auth/github/callback"
-
-        def _work():
-            tok = httpx.post("https://github.com/login/oauth/access_token", data={
-                "client_id": config.github_login_client,
-                "client_secret": config.github_login_secret,
-                "code": code, "redirect_uri": redirect,
-            }, headers={"Accept": "application/json"}, timeout=15).json()
-            at = tok.get("access_token")
-            u = httpx.get("https://api.github.com/user", headers={
-                "Authorization": "Bearer " + (at or ""),
-                "Accept": "application/vnd.github+json"}, timeout=15).json()
-            return u.get("login") or ""
-        try:
-            login = await asyncio.to_thread(_work)
-        except Exception as exc:
-            return _login_denied_html(f"Falha no login GitHub: {exc}")
-        if not login:
-            return _login_denied_html("Não consegui obter seu usuário do GitHub.")
-        allowed = memory.get_setting("login_github_user")
-        if not allowed:
-            memory.set_setting("login_github_user", login)  # pin the first login
-        elif login.lower() != allowed.lower():
-            return _login_denied_html("Este usuário GitHub não tem acesso a esta E.V.")
-        return _login_ok_html()
 
     return app
